// src/app/shared/models/guardianinfo.ts
export class Guardianinfo {
    id: number;
    salutation: string;
    fullname: string;
    emailid: string;
    relationship: string;
    occupation: string;
    cellphone: string;
    alternatephone: string;
    streetno: string;
    streetname: string;
    city: string;
    postalcode: string;
    state: string;
    country: string;
    constructor(){

    }
}